requirements.txt
certifi==2019.9.11
chardet==3.0.4
click==7.1.2
Flask==1.1.2
gunicorn==20.0.4
idna==2.8
itsdangerous==1.1.0
Jinja2==2.11.2
MarkupSafe==1.1.1
PyJWT==1.7.1
pytz==2019.3
requests==2.22.0
six==1.12.0
urllib3==1.25.6
Werkzeug==1.0.1
Tensorflow-cpu==2.3.0
pillow